fix but also remove getStaticTypes
This was set up to read types based on ts files in this repo, so I modified it to read types from material-ui. Unfortunately, that results in a huge number of components and props that takes too long to parse during a build. Plus there are way too many useless props (HTML attributes, system props, etc). I need to rethink my approach to this, and it will make the most sense to do that rethinking alongside figuring out snippet generation.